# payment-service/src/main/java/com/solmaz/paymentservice/model/PaymentRequest.java
package com.solmaz.paymentservice.model;

public class PaymentRequest {

    private Integer bookingId;
    private String paymentType;

    public Integer getBookingId() {
        return bookingId;
    }

    public void setBookingId(Integer bookingId) {
        this.bookingId = bookingId;
    }

    public String getPaymentType() {
        return paymentType;
    }

    public void setPaymentType(String paymentType) {
        this.paymentType = paymentType;
    }
}