# ticket-planner-main-service/src/main/java/com/solmaz/ticketplannermainservice/dto/response/UserResponse.java
package com.solmaz.ticketplannermainservice.dto.response;

public class UserResponse {

    private String firstName;
    private String middleName;
    private String lastName;
    private String email;
    private String telephoneNumber;

    public UserResponse() {
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getMiddleName() {
        return middleName;
    }

    public void setMiddleName(String middleName) {
        this.middleName = middleName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getTelephoneNumber() {
        return telephoneNumber;
    }

    public void setTelephoneNumber(String telephoneNumber) {
        this.telephoneNumber = telephoneNumber;
    }
}